requirements.txt
mysql-connector-python==9.2.0
python-dotenv==1.0.1
streamlit==1.42.1
streamlit-authenticator==0.4.2
streamlit-option-menu==0.4.0